gacp

+++ Commands/Dic_menu.py
@@ -6,20 +6,24 @@ for k in CShowFile.keys():
     CShowFile[k] = False
 
 
-def Default_Values(Q,project_path,parent_keys=[]):
+def Default_Values(Q,project_path,parent_keys=[],read_only=False):
     D = {}
     os.system('mkdir -p '+opj(project_path,'__local__'))
     D['project_path'] = project_path
     D['.pkl'] = opj(project_path,'__local__','default_values.pkl')
-    os.system('rm '+D['.pkl'])
+    
     D['parent_keys'] = parent_keys
     D['current_keys'] = ['Q']
     D['Q'] = Q
+    D['read_only'] = read_only
+    if not read_only:
+        os.system('rm '+D['.pkl'])
     add_keys(D,D['Q'])
     def function_save():
+        if D['read_only']:
+            return
         save_C(D['Q'],D['project_path'])
     def function_load():
-        #cm(-1,ra=1)
         load_C(D['Q'],D['project_path'])
         add_keys(D,D['Q'])
     def function_up():
@@ -49,27 +53,29 @@ def Default_Values(Q,project_path,parent_keys=[]):
             } 
 
     def function_show(message=''):
+        clear_screen()
+        print D['project_path'].replace(opjh(),'')
         return show_menu(key_access(D,D['current_keys']),message)
 
     def function_menu():
-        
         message = ''
+        if not D['read_only']:
+            D['save']()
         while True:
             D['load']()
             items = D['show'](message)
             while True:
-                R = choice(items)
+                R = make_choice(items)
                 if R['message'] != 'ran python':
                     break
-            #cb('R =',R)
             if R['action'] == 'quit':
                 cw("\ndone.\n")
-                return#sys.exit()
+                return #sys.exit()
             elif R['action'] == 'failure':
                 message = R['message']
                 continue
             elif R['action'] == 'continue':
-                message = 'continue'
+                message = R['message'] #'continue'
                 continue
             elif R['action'] == 'choose':
                 
@@ -81,13 +87,11 @@ def Default_Values(Q,project_path,parent_keys=[]):
                     message = 'already at <top>'
                 else:
                     S = D['down'](R['message'])
-                    #cb('S =',S)
                     message = S['message']
                     if S['action'] == 'success':
                         continue
                     else:
                         K = key_access(D,D['current_keys'])
-                        #message = d2s("need to set",R['message'],'(',K[R['message']],')')
                         message = set_value(D,R['message'])
                         continue
 
@@ -112,7 +116,15 @@ def set_value(D,key):
         raw_enter()
         message = K[key]
         return message
-    value = input(d2n("Enter value for '",key,"' (",K[key],"): "))
+    if type(K[key]) == bool:
+        yes_no = raw_input(d2n("Toggle '",key,"'? ([y]/n) "))
+        if yes_no == 'y' or yes_no == '':
+            value = not K[key]
+        else:
+            message = d2n("'",key,"' unchanged")
+            return message
+    else:
+        value = input(d2n("Enter value for '",key,"' (",K[key],"): "))
     if type(value) != type(K[key]):
         message = d2n("*** type(",value,") != type(",K[key],")")
     else:
@@ -139,27 +151,28 @@ def key_access(Dic,keys,start=True):
 def add_keys(D,E,keys=[]):
     E['--keys--'] = keys
     if '--mode--' not in E:
-        E['--mode--'] = 'var'
+        if D['read_only']:
+            E['--mode--'] = 'const'
+        else:
+            E['--mode--'] = 'var'
+    elif D['read_only'] and E['--mode--'] == 'var':
+        E['--mode--'] = 'const'
+
     for k in E.keys():
         if type(E[k]) == dict:
             add_keys(D,E[k],keys+[k])
 
 
 def load_C(C,project_path,name='default_values'):
-    #cm(0,ra=1)
     if len(sggo(opj(project_path,'__local__','ready'))) == 0:
         return False
     if len(sggo(opj(project_path,'__local__',name+'.writing.pkl'))) > 0:
         return False
     if len(sggo(opj(project_path,'__local__',name+'.pkl'))) == 0:
         return False
-    #cm(1,ra=1)
     try:
-        #cm(2)
         D = lo(opj(project_path,'__local__',name+'.pkl'))
-        #cm(3)
     except:
-        #cm(4)
         return False
     if len(sggo(opj(project_path,'__local__','ready'))) == 0:
         return False
@@ -205,14 +218,12 @@ def clear_screen():
     cr("\nclear screen\n")
 
 def show_menu(C,message):
-    clear_screen()
     if '--keys--' in C:
         key_list = C['--keys--']
     else:
         key_list = ['no keys']
-    #cg(key_list,ra=1)
     cprint(
-        '/'.join(key_list),
+        '/'+'/'.join(key_list),
         attrs=['bold','reverse'],
         color='white',
         on_color='on_blue'
@@ -227,12 +238,11 @@ def show_menu(C,message):
     else:
         s = '<up>'
     sorted_keys.insert(0,s)
-    #cy(sorted_keys,ra=1)
     cc = bl
     for i in rlen(sorted_keys):
         if i == 0 and s == '<top>':
             cb(s)
-            continue#return ''
+            continue #return ''
         k = sorted_keys[i]
         val_color = lb
         v = ''
@@ -243,7 +253,8 @@ def show_menu(C,message):
                 for l in C[k]:
                     if not is_meta_key(l):
                         ctr += 1
-                v = d2n(ctr)
+                #v = d2n(ctr)
+                v = ''
                 cc = wh+underlined
                 val_color = wh
             elif is_meta_key(k):
@@ -262,7 +273,6 @@ def show_menu(C,message):
 
         cb(bl,i,cc+k+colored.attr('res_underlined'),val_color,v,edited)
     cg('$',message)
-    #print sorted_keys
     return sorted_keys
 
 
@@ -298,15 +308,13 @@ def Dic_Loader(path,wait_time=0.2):
 
 
 
-
-
-def choice(sorted_keys):
+def make_choice(sorted_keys):
 
     raw_choice = raw_input(wh+'choice: '+lb)
 
     if raw_choice == '':
         return {
-            'message':"other commands: 'load','save','q' (quit), 'p' (python)",
+            'message':"other commands: 'q' (quit), 'p' (python)",
             'action':'continue',
         }
         
@@ -318,7 +326,6 @@ def choice(sorted_keys):
 
     if raw_choice == 'p':
         menu_python()
-        #return choice(sorted_keys)
         return {
             'message':  "ran python",
             'action':   'continue',
@@ -353,7 +360,6 @@ def menu_python():
             d[-1]=d2s('print(',d[-1],')')
         code = ';'.join(d)
         exec(code)
-        #raw_enter()
         return ''
     except Exception as e:
         exc_type, exc_obj, exc_tb = sys.exc_info()
@@ -362,10 +368,45 @@ def menu_python():
         return message
 
 
+if __name__ == '__main__':
 
-from kzpy3.Commands.temp3 import Q
-D = Default_Values(Q,opjk('Commands'))
-D['menu']()
+    Default_Args = {
+        'menu':True,
+        'read_only':False,
+        'dic_project_path':opjk('Commands/A'),
+    }
+
+    for d in Default_Args:
+        if d not in Arguments:
+            Arguments[d] = Default_Args[d]
+
+
+
+    #dic_path = Arguments['dic_path']
+    #dic_project_path = '/'.join(dic_path.split('/')[:-1])
+    dic_project_path = Arguments['dic_project_path']
+
+    exec_str = d2s(
+        'from',
+        project_path__to__project_import_prefix(dic_project_path)+'.default_values',
+        'import Q',
+    )
+    exec(exec_str)
+    D = Default_Values(Q,dic_project_path,read_only=Arguments['read_only'])
+
+    if False:
+        from kzpy3.Commands.temp3 import Q
+        D = Default_Values(Q,opjk('Commands'),read_only=Arguments['read_only'])
+
+    if Arguments['read_only']:
+        D['load']()
+
+    if Arguments['menu']:
+        D['menu']()
+
+
+
+
 
 #EOF
